requirements.txt
wikipedia-api
ollama
numpy
scikit-learn 
faiss-cpu
chromadb
sentence-transformers
fuzzywuzzy
wikipedia
langchain
nltk
